requirements.txt
Flask
apscheduler
websocket-client
gunicorn
pandas
psycopg2
scipy
scikit-learn
sklearn
numpy
redis
requests
